backend/requirements.txt
Django==3.2
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
gunicorn==20.1.0
isort==5.10.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
